+++ yike-20260707/src/Models/ListInfiniteCanvasesRequest.php
@@ -26,18 +26,30 @@ class ListInfiniteCanvasesRequest extends Model
     /**
      * @var string
      */
+    public $productionId;
+
+    /**
+     * @var string
+     */
     public $sortBy;
 
     /**
      * @var string
      */
     public $sortOrder;
+
+    /**
+     * @var string
+     */
+    public $workspaceId;
     protected $_name = [
         'keyword' => 'Keyword',
         'pageNo' => 'PageNo',
         'pageSize' => 'PageSize',
+        'productionId' => 'ProductionId',
         'sortBy' => 'SortBy',
         'sortOrder' => 'SortOrder',
+        'workspaceId' => 'WorkspaceId',
     ];
 
     public function validate()
@@ -60,12 +72,20 @@ class ListInfiniteCanvasesRequest extends Model
             $res['PageSize'] = $this->pageSize;
         }
 
+        if (null !== $this->productionId) {
+            $res['ProductionId'] = $this->productionId;
+        }
+
         if (null !== $this->sortBy) {
             $res['SortBy'] = $this->sortBy;
         }
 
         if (null !== $this->sortOrder) {
             $res['SortOrder'] = $this->sortOrder;
+        }
+
+        if (null !== $this->workspaceId) {
+            $res['WorkspaceId'] = $this->workspaceId;
         }
 
         return $res;
@@ -91,12 +111,20 @@ class ListInfiniteCanvasesRequest extends Model
             $model->pageSize = $map['PageSize'];
         }
 
+        if (isset($map['ProductionId'])) {
+            $model->productionId = $map['ProductionId'];
+        }
+
         if (isset($map['SortBy'])) {
             $model->sortBy = $map['SortBy'];
         }
 
         if (isset($map['SortOrder'])) {
             $model->sortOrder = $map['SortOrder'];
+        }
+
+        if (isset($map['WorkspaceId'])) {
+            $model->workspaceId = $map['WorkspaceId'];
         }
 
         return $model;